QBWROneSingleStacks in own file now and exported
